requirements.txt
asgiref==3.2.10
astroid==2.4.2
dj-database-url==0.5.0
Django==3.1.2
django-cors-headers==3.5.0
django-crispy-forms==1.10.0
django-heroku==0.3.1
djangorestframework==3.12.1
ecdsa==0.14.1
gunicorn==20.0.4
isort==5.6.4
lazy-object-proxy==1.4.3
mccabe==0.6.1
Pillow==8.0.1
psycopg2==2.8.6
pyasn1==0.4.8
pycparser==2.20
pylint==2.6.0
python-jose==3.2.0
pytz==2020.1
rsa==4.6
six==1.15.0
sqlparse==0.4.1
toml==0.10.1
whitenoise==5.2.0
wrapt==1.12.1
